requirements.txt
torch
torchvision
pillow
flask
flask_cors